requirements.txt
# scrapy==2.0.1
pandas==0.25.0
ipdb==0.12.2
google-api-python-client==1.7.11
google-auth-oauthlib==0.3.0